+++ src/uet/oop/bomberman/entities/Grass.java
@@ -4,7 +4,7 @@ import javafx.scene.image.Image;
 
 public class Grass extends Entity {
 
-    public Grass(int x, int y, Image img) {
+    public Grass(double x, double y, Image img) {
         super(x, y, img);
     }
